src/main/java/org/example/MainTest.java
package org.example;


import org.junit.jupiter.api.AfterEach;
import org.junit.jupiter.api.BeforeEach;
import org.junit.jupiter.api.Test;
import org.mockito.Mockito;

import java.io.ByteArrayInputStream;
import java.io.InputStream;
import java.util.Scanner;

public class MainTest {

    private final InputStream systemIn = System.in;
    private ByteArrayInputStream testIn;

    @BeforeEach
    public void setUp() {
        testIn = new ByteArrayInputStream("1\n3\n2\n5\n1 2 3 4 5\n".getBytes());
        System.setIn(testIn);
    }

    @AfterEach
    public void tearDown() {
        System.setIn(systemIn);
        testIn = null;
    }

    @Test
    public void testAutoBookSeats() {
        Cinema cinema = Mockito.mock(Cinema.class);
        Scanner scanner = new Scanner(System.in);
        Main.autoBookSeats(cinema, scanner);
        // Verify or assert any expected behavior
    }

    @Test
    public void testBookSeats() {
        Cinema cinema = Mockito.mock(Cinema.class);
        Scanner scanner = new Scanner(System.in);
        Main.bookSeats(cinema, scanner);
        // Verify or assert any expected behavior
    }

    @Test
    public void testCancelBooking() {
        Cinema cinema = Mockito.mock(Cinema.class);
        Scanner scanner = new Scanner(System.in);
        Main.cancelBooking(cinema, scanner);
        // Verify or assert any expected behavior
    }

    @Test
    public void testCheckAvailability() {
        Cinema cinema = Mockito.mock(Cinema.class);
        Scanner scanner = new Scanner(System.in);
        Main.checkAvailability(cinema, scanner);
        // Verify or assert any expected behavior
    }

    @Test
    public void testPrintSeatingArrangement() {
        Cinema cinema = Mockito.mock(Cinema.class);
        Scanner scanner = new Scanner(System.in);
        Main.printSeatingArrangement(cinema, scanner);
        // Verify or assert any expected behavior
    }
}